# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kabadiwala Connect API",
    "description": "Vernacular, low-literacy offline-tolerant platform connecting informal e-waste collectors with authorized recyclers.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recyclers": {
      "get": {
        "tags": [
          "Recyclers"
        ],
        "summary": "List authorized recyclers",
        "operationId": "list_recyclers_api_v1_recyclers_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by authorization status",
              "title": "Status"
            },
            "description": "Filter by authorization status"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by accepted category",
              "title": "Category"
            },
            "description": "Filter by accepted category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecyclerResponse"
                  },
                  "title": "Response List Recyclers Api V1 Recyclers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recyclers"
        ],
        "summary": "Register a new recycler facility",
        "operationId": "create_recycler_api_v1_recyclers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecyclerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecyclerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recyclers/{recycler_id}": {
      "get": {
        "tags": [
          "Recyclers"
        ],
        "summary": "Get recycler by ID",
        "operationId": "get_recycler_api_v1_recyclers__recycler_id__get",
        "parameters": [
          {
            "name": "recycler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recycler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecyclerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recyclers"
        ],
        "summary": "Update recycler details",
        "operationId": "update_recycler_api_v1_recyclers__recycler_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recycler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recycler Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecyclerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecyclerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recyclers/{recycler_id}/verify": {
      "post": {
        "tags": [
          "Recyclers"
        ],
        "summary": "Admin approves & verifies recycler authorization",
        "operationId": "verify_recycler_api_v1_recyclers__recycler_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recycler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recycler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecyclerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recyclers/{recycler_id}/suspend": {
      "post": {
        "tags": [
          "Recyclers"
        ],
        "summary": "Admin suspends recycler authorization",
        "operationId": "suspend_recycler_api_v1_recyclers__recycler_id__suspend_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recycler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recycler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecyclerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recyclers/jobs/audit-expired": {
      "post": {
        "tags": [
          "Recyclers"
        ],
        "summary": "Audit job: auto-flag recyclers with expired EPR authorization",
        "operationId": "audit_expired_authorizations_api_v1_recyclers_jobs_audit_expired_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/sync/push": {
      "post": {
        "tags": [
          "Offline Synchronization"
        ],
        "summary": "Batch upload offline client operations (Idempotent)",
        "operationId": "sync_push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/pull": {
      "get": {
        "tags": [
          "Offline Synchronization"
        ],
        "summary": "Delta download updated master data, prices, recyclers, and collector records",
        "operationId": "sync_pull_api_v1_sync_pull_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 timestamp cursor",
              "title": "Since"
            },
            "description": "ISO-8601 timestamp cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "sub_category": {
            "type": "string",
            "title": "Sub Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "image_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Ref"
          },
          "approx_weight_kg": {
            "type": "number",
            "title": "Approx Weight Kg"
          },
          "condition": {
            "type": "string",
            "title": "Condition"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "estimated_value": {
            "type": "number",
            "title": "Estimated Value"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "sub_category",
          "description",
          "image_ref",
          "approx_weight_kg",
          "condition",
          "source_type",
          "estimated_value",
          "created_at",
          "updated_at"
        ],
        "title": "MaterialResponse"
      },
...
      "PriceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "sub_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Category"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "buying_price": {
            "type": "number",
            "title": "Buying Price"
          },
          "selling_quoted_price": {
            "type": "number",
            "title": "Selling Quoted Price"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "market_min": {
            "type": "number",
            "title": "Market Min"
          },
          "market_max": {
            "type": "number",
            "title": "Market Max"
          },
          "recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recycler Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "sub_category",
          "district",
          "city",
          "latitude",
          "longitude",
          "recorded_at",
          "buying_price",
          "selling_quoted_price",
          "unit",
          "market_min",
          "market_max",
          "recycler_id",
          "source",
          "created_at"
        ],
        "title": "PriceResponse"
      },
      "RecyclerCreate": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 36,
            "title": "Id",
            "description": "e.g. REC-MH-042"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "materials_accepted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Materials Accepted"
          },
          "authorization_number": {
            "type": "string",
            "maxLength": 100,
            "title": "Authorization Number"
          },
          "authorization_body": {
            "type": "string",
            "pattern": "^(CPCB|SPCB|OTHER)$",
            "title": "Authorization Body",
            "default": "CPCB"
          },
          "authorization_valid_till": {
            "type": "string",
            "format": "date",
            "title": "Authorization Valid Till"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "title": "Phone"
          },
          "offered_rates": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Offered Rates"
          },
          "pickup_available": {
            "type": "boolean",
            "title": "Pickup Available",
            "default": false
          },
          "pickup_radius_km": {
            "type": "number",
            "title": "Pickup Radius Km",
            "default": 0.0
          },
          "service_area": {
            "additionalProperties": true,
            "type": "object",
            "title": "Service Area"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "latitude",
          "longitude",
          "authorization_number",
          "authorization_valid_till",
          "phone"
        ],
        "title": "RecyclerCreate"
      },
      "RecyclerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "materials_accepted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Materials Accepted"
          },
          "authorization_number": {
            "type": "string",
            "title": "Authorization Number"
          },
          "authorization_body": {
            "type": "string",
            "title": "Authorization Body"
          },
          "authorization_status": {
            "type": "string",
            "title": "Authorization Status"
          },
          "authorization_valid_till": {
            "type": "string",
            "format": "date",
            "title": "Authorization Valid Till"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "offered_rates": {
            "additionalProperties": true,
            "type": "object",
            "title": "Offered Rates"
          },
          "pickup_available": {
            "type": "boolean",
            "title": "Pickup Available"
          },
          "pickup_radius_km": {
            "type": "number",
            "title": "Pickup Radius Km"
          },
          "service_area": {
            "additionalProperties": true,
            "type": "object",
            "title": "Service Area"
          },
          "rating": {
            "type": "number",
            "title": "Rating"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "latitude",
          "longitude",
          "materials_accepted",
          "authorization_number",
          "authorization_body",
          "authorization_status",
          "authorization_valid_till",
          "phone",
          "offered_rates",
          "pickup_available",
          "pickup_radius_km",
          "service_area",
          "rating",
          "created_at",
          "updated_at"
        ],
        "title": "RecyclerResponse"
      },
      "RecyclerUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "materials_accepted": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materials Accepted"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "offered_rates": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offered Rates"
          },
          "pickup_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Available"
          },
          "pickup_radius_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Radius Km"
          },
          "service_area": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Area"
          }
        },
        "type": "object",
        "title": "RecyclerUpdate"
      },
      "SafetyContentItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "hazard_level": {
            "type": "string",
            "title": "Hazard Level"
          },
          "pictogram_url": {
            "type": "string",
            "title": "Pictogram Url"
          },
          "audio_prompt_urls": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Audio Prompt Urls"
          },
          "title_vernacular": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Title Vernacular"
          },
          "instructions_vernacular": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Instructions Vernacular"
          },
          "dos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dos"
          },
          "donts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Donts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "hazard_level",
          "pictogram_url",
          "audio_prompt_urls",
          "title_vernacular",
          "instructions_vernacular",
          "dos",
          "donts"
        ],
        "title": "SafetyContentItem"
      },
      "SyncOperation": {
        "properties": {
          "client_tx_id": {
            "type": "string",
            "title": "Client Tx Id",
            "description": "Unique client transaction UUID for idempotency"
          },
          "action": {
            "type": "string",
            "pattern": "^(create_lot|update_lot|record_cash_payment)$",
            "title": "Action"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "client_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Client Timestamp"
          }
        },
        "type": "object",
        "required": [
          "client_tx_id",
          "action",
          "payload",
          "client_timestamp"
        ],
        "title": "SyncOperation"
      },
      "SyncPullResponse": {
        "properties": {
          "cursor": {
            "type": "string",
            "title": "Cursor"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialResponse"
            },
            "type": "array",
            "title": "Materials"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PriceResponse"
            },
            "type": "array",
            "title": "Prices"
          },
          "recyclers": {
            "items": {
              "$ref": "#/components/schemas/RecyclerResponse"
            },
            "type": "array",
            "title": "Recyclers"
          },
          "safety_content": {
            "items": {
              "$ref": "#/components/schemas/SafetyContentItem"
            },
            "type": "array",
            "title": "Safety Content"
          },
          "my_lots": {
            "items": {
              "$ref": "#/components/schemas/LotResponse"
            },
            "type": "array",
            "title": "My Lots"
          },
          "my_ledger": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryResponse"
            },
            "type": "array",
            "title": "My Ledger"
          }
        },
        "type": "object",
        "required": [
          "cursor",
          "server_time",
          "materials",
          "prices",
          "recyclers",
          "safety_content",
          "my_lots",
          "my_ledger"
        ],
        "title": "SyncPullResponse"
      },
      "SyncPushRequest": {
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/SyncOperation"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Operations"
          }
        },
        "type": "object",
        "required": [
          "operations"
        ],
        "title": "SyncPushRequest"
      },
      "SyncPushResponse": {
        "properties": {
          "processed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Processed",
            "description": "List of processed client_tx_ids"
          },
          "conflicts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "processed",
          "server_time"
        ],
        "title": "SyncPushResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}